Create App type
