+++ packages/web-components/src/index.ts
@@ -6,3 +6,4 @@ export *  from './initialize';
 export *  from './defaultProductProperties';
 export *  from './tracking';
 export *  from './updateContextSettings';
+export *  from './app';
